Adding in bits needed to clean up when people leave channel
SoundChan will leave a channel after 10 seconds if it is the only member
left. When it leaves it will empty the queue of audio to play.

When SoundChan is the last member in a channel, it will pause any
playing audio.

New command to drop play queue and stop current audio.